base-image: Remove the pip & virtualenv cache
Pip stores cache in `~/.cache/pip/` and when the
virtualenv is invoked there is also a cache
created in `~/.local/share/virtualenv/`. This
cache is not required for the base image and can
be removed to reduce the size of the image. Also
if this cache has older / vulnerable software they
are flagged by the security scanners.

Signed-off-by: Suraj Deshmukh <suraj.deshmukh@microsoft.com>